rename Plan to PlanEntity and refactor fields
