+++ lib/models/map_route_state.dart
@@ -1,5 +1,6 @@
 import 'package:async/async.dart';
 import 'package:equatable/equatable.dart';
+import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
 import 'package:trufi_core/trufi_models.dart';
 
 class MapRouteState extends Equatable {
@@ -21,7 +22,7 @@ class MapRouteState extends Equatable {
 
   final TrufiLocation fromPlace;
   final TrufiLocation toPlace;
-  final Plan plan;
+  final PlanEntity plan;
   final Ad ad;
   final bool isFetching;
   final bool showSuccessAnimation;
@@ -29,11 +30,11 @@ class MapRouteState extends Equatable {
   MapRouteState copyWith({
     TrufiLocation fromPlace,
     TrufiLocation toPlace,
-    Plan plan,
+    PlanEntity plan,
     Ad ad,
     bool isFetching,
     bool showSuccessAnimation,
-    CancelableOperation<Plan> currentFetchPlanOperation,
+    CancelableOperation<PlanEntity> currentFetchPlanOperation,
     CancelableOperation<Ad> currentFetchAdOperation,
   }) {
     return MapRouteState(
@@ -52,7 +53,7 @@ class MapRouteState extends Equatable {
       fromPlace:
           TrufiLocation.fromJson(json[_fromPlace] as Map<String, dynamic>),
       toPlace: TrufiLocation.fromJson(json[_toPlace] as Map<String, dynamic>),
-      plan: Plan.fromJson(json[_plan] as Map<String, dynamic>),
+      plan: PlanEntity.fromJson(json[_plan] as Map<String, dynamic>),
       ad: Ad.fromJson(json[_ad] as Map<String, dynamic>),
       isFetching: json[_isFetching] as bool ?? false,
       showSuccessAnimation: json[_showSuccessAnimation] as bool ?? false,
