+++ lib/plan/plan.dart
@@ -2,6 +2,7 @@ import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/material.dart';
 import 'package:rxdart/rxdart.dart' as rx;
 import 'package:trufi_core/blocs/home_page_cubit.dart';
+import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
 import 'package:trufi_core/trufi_configuration.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -9,7 +10,6 @@ import '../composite_subscription.dart';
 import '../plan/plan_itinerary_tabs.dart';
 import '../plan/plan_map.dart';
 import '../trufi_app.dart';
-import '../trufi_models.dart';
 
 class PlanPageController {
   PlanPageController(this.plan, this.ad) {
@@ -20,7 +20,7 @@ class PlanPageController {
     );
   }
 
-  final Plan plan;
+  final PlanEntity plan;
   final Ad ad;
 
   final _selectedItineraryController = rx.BehaviorSubject<PlanItinerary>();
@@ -45,7 +45,7 @@ class PlanPageController {
 }
 
 class PlanPage extends StatefulWidget {
-  final Plan plan;
+  final PlanEntity plan;
   final Ad ad;
   final LocaleWidgetBuilder customOverlayWidget;
   final WidgetBuilder customBetweenFabWidget;
