+++ lib/services/plan_request/online_repository.dart
@@ -6,6 +6,7 @@ import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
 import 'package:package_info/package_info.dart';
+import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
 import 'package:trufi_core/l10n/trufi_localization.dart';
 import 'package:trufi_core/repository/exception/fetch_online_exception.dart';
 import 'package:trufi_core/trufi_models.dart';
@@ -21,7 +22,7 @@ class OnlineRepository implements RequestManager {
   OnlineRepository({@required this.otpEndpoint, this.adsEndpoint});
 
   @override
-  CancelableOperation<Plan> fetchTransitPlan(
+  CancelableOperation<PlanEntity> fetchTransitPlan(
     TrufiLocation from,
     TrufiLocation to,
     String correlationId,
@@ -30,7 +31,7 @@ class OnlineRepository implements RequestManager {
   }
 
   @override
-  CancelableOperation<Plan> fetchCarPlan(
+  CancelableOperation<PlanEntity> fetchCarPlan(
     TrufiLocation from,
     TrufiLocation to,
     String correlationId,
@@ -46,14 +47,14 @@ class OnlineRepository implements RequestManager {
     return _fetchCancelableAd(to, correlationId);
   }
 
-  CancelableOperation<Plan> _fetchCancelablePlan(
+  CancelableOperation<PlanEntity> _fetchCancelablePlan(
     TrufiLocation from,
     TrufiLocation to,
     String mode,
     String correlationId,
   ) {
     return CancelableOperation.fromFuture(() async {
-      final Plan plan = await _fetchPlan(from, to, mode, correlationId);
+      final PlanEntity plan = await _fetchPlan(from, to, mode, correlationId);
       return plan;
     }());
   }
@@ -68,7 +69,7 @@ class OnlineRepository implements RequestManager {
     }());
   }
 
-  Future<Plan> _fetchPlan(
+  Future<PlanEntity> _fetchPlan(
     TrufiLocation from,
     TrufiLocation to,
     String mode,
@@ -188,8 +189,8 @@ List<TrufiLocation> _parseLocations(String responseBody) {
       .toList() as List<TrufiLocation>;
 }
 
-Plan _parsePlan(String responseBody) {
-  return Plan.fromJson(json.decode(responseBody) as Map<String, dynamic>);
+PlanEntity _parsePlan(String responseBody) {
+  return PlanEntity.fromJson(json.decode(responseBody) as Map<String, dynamic>);
 }
 
 Ad _parseAd(String responseBody) {
