+++ lib/services/plan_request/request_manager.dart
@@ -1,14 +1,15 @@
 import 'package:async/async.dart';
+import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
 import 'package:trufi_core/trufi_models.dart';
 
 abstract class RequestManager {
-  CancelableOperation<Plan> fetchTransitPlan(
+  CancelableOperation<PlanEntity> fetchTransitPlan(
     TrufiLocation from,
     TrufiLocation to,
     String correlationId,
   );
 
-  CancelableOperation<Plan> fetchCarPlan(
+  CancelableOperation<PlanEntity> fetchCarPlan(
     TrufiLocation from,
     TrufiLocation to,
     String correlationId,
