+++ test/trufi_map_util_unit_test.dart
@@ -3,9 +3,8 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:latlong/latlong.dart';
 
+import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
 import 'package:trufi_core/trufi_map_utils.dart';
-import 'package:trufi_core/trufi_models.dart';
-
 void main() {
   const double _latitude = 18.0;
   const double _longitude = 129.0;
