+++ test/trufi_models_trufi_location_unit_test.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:latlong/latlong.dart';
+import 'package:trufi_core/entities/plan_entities/plan_entity.dart';
 
 import 'package:trufi_core/trufi_models.dart';
